docs(refactor/feat): extract readme functions etc into separate file
Because of the amount of work that we need to do to treat the readmes,
it made sense to move it into a separate module. More manageable that way
